Create CaptchaGenerationTask
