+++ proxy/src/main/java/ru/leymooo/botfilter/captcha/CaptchaGeneration.java
@@ -1,29 +1,22 @@
 package ru.leymooo.botfilter.captcha;
 
-import java.awt.Color;
+import com.google.common.util.concurrent.ThreadFactoryBuilder;
 import java.awt.Font;
-import java.awt.image.BufferedImage;
 import java.util.ArrayList;
 import java.util.Arrays;
 import java.util.Collections;
 import java.util.List;
-import java.util.Random;
 import java.util.concurrent.ExecutorService;
 import java.util.concurrent.Executors;
-import java.util.concurrent.ThreadFactory;
-import java.util.concurrent.ThreadLocalRandom;
 import java.util.concurrent.ThreadPoolExecutor;
-import java.util.concurrent.atomic.AtomicInteger;
 import java.util.logging.Level;
 import lombok.experimental.UtilityClass;
 import net.md_5.bungee.BungeeCord;
 import ru.leymooo.botfilter.caching.CachedCaptcha;
 import ru.leymooo.botfilter.caching.PacketUtils;
 import ru.leymooo.botfilter.captcha.generator.CaptchaPainter;
-import ru.leymooo.botfilter.captcha.generator.map.CraftMapCanvas;
 import ru.leymooo.botfilter.captcha.generator.map.MapPalette;
 import ru.leymooo.botfilter.config.Settings;
-import ru.leymooo.botfilter.packets.MapDataPacket;
 
 /**
  * @author Leymooo
@@ -42,7 +35,7 @@ public class CaptchaGeneration
 
         generation = true;
         Thread thread = new Thread( CaptchaGeneration::generateCaptchas );
-        thread.setName( "CaptchaGeneration-provider-thread" );
+        thread.setName( "CaptchaGenerationProvider-thread" );
         thread.setPriority( Thread.MIN_PRIORITY );
         thread.start();
     }
@@ -58,7 +51,10 @@ public class CaptchaGeneration
             PacketUtils.captchas.clear();
             BungeeCord.getInstance().getLogger().log( Level.INFO, "[BotFilter] " + ( BungeeCord.getInstance().isEnabled() ? "Начата генерация капчи в фоне." : "Генерация капчи продолжится параллельно с загрузкой BungeeCord." ) );
             ExecutorService executor = Executors.newFixedThreadPool( Runtime.getRuntime().availableProcessors(),
-                    new CaptchaThreadFactory() );
+                    new ThreadFactoryBuilder()
+                            .setPriority( Thread.MIN_PRIORITY )
+                            .setNameFormat( "CaptchaGenerationTask-thread-%d" )
+                            .build() );
             CaptchaPainter painter = new CaptchaPainter();
             MapPalette.prepareColors();
 
@@ -73,7 +69,7 @@ public class CaptchaGeneration
 
             for ( int i = 1; i <= captchaCount; i++ )
             {
-                executor.execute( () -> generate( executor, painter, fonts, holders ) );
+                executor.execute( new CaptchaGenerationTask( executor, painter, fonts, holders ) );
             }
 
             long start = System.currentTimeMillis();
@@ -111,85 +107,6 @@ public class CaptchaGeneration
         } finally
         {
             generation = false;
-        }
-    }
-
-    private static void generate(ExecutorService executor, CaptchaPainter painter, List<Font> fonts,
-                                 List<CachedCaptcha.CaptchaHolder> holders)
-    {
-        try
-        {
-            Random rnd = ThreadLocalRandom.current();
-            String answer = randomAnswer( rnd );
-            BufferedImage image = painter.draw( fonts.get( rnd.nextInt( fonts.size() ) ), randomNotWhiteColor( rnd ), answer );
-            final CraftMapCanvas map = new CraftMapCanvas();
-            map.drawImage( 0, 0, image );
-            MapDataPacket packet = new MapDataPacket( 0, (byte) 0, map.getMapData() );
-            CachedCaptcha.CaptchaHolder holder = CachedCaptcha.createCaptchaPacket( packet, answer );
-            holders.add( holder );
-        } catch ( Throwable e )
-        {
-            //Прекращаем генерацию если случилась любая ошибка
-            e.printStackTrace();
-            executor.shutdownNow();
-        }
-    }
-
-    private static Color randomNotWhiteColor(Random rnd)
-    {
-        Color color = MapPalette.colors[rnd.nextInt( MapPalette.colors.length )];
-
-        int r = color.getRed();
-        int g = color.getGreen();
-        int b = color.getBlue();
-
-        if ( r == 255 && g == 255 && b == 255 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        if ( r == 220 && g == 220 && b == 220 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        if ( r == 199 && g == 199 && b == 199 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        if ( r == 255 && g == 252 && b == 245 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        if ( r == 220 && g == 217 && b == 211 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        if ( r == 247 && g == 233 && b == 163 )
-        {
-            return randomNotWhiteColor( rnd );
-        }
-        return color;
-    }
-    private static String randomAnswer(Random rnd)
-    {
-        if ( rnd.nextBoolean() )
-        {
-            return Integer.toString( rnd.nextInt( ( 99999 - 10000 ) + 1 ) + 10000 );
-        } else
-        {
-            return Integer.toString( rnd.nextInt( ( 9999 - 1000 ) + 1 ) + 1000 );
-        }
-    }
-    //Для правильного именования потоков которые генерируют капчу и для создания низкого приоритета
-    private static class CaptchaThreadFactory implements ThreadFactory
-    {
-        private static final AtomicInteger counter = new AtomicInteger();
-        @Override
-        public Thread newThread(Runnable task)
-        {
-            Thread thr = new Thread( task );
-            thr.setPriority( Thread.MIN_PRIORITY );
-            thr.setName( "CaptchaGenerator-thread-" + counter.incrementAndGet() );
-            return thr;
         }
     }
 }
